Split files to fix cycle
